% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\docType{data}
\name{Calibri}
\alias{Calibri}
\alias{font_cal}
\alias{font_cal_light}
\title{Calibri Font variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_cal

font_cal_light
}
\description{
font_cal == "Calibri"

font_cal_light == "Calibri Light"
}
\keyword{datasets}
