% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pwd.R
\name{ggplot_add.pwd_theme}
\alias{ggplot_add.pwd_theme}
\title{Add PWD themes to ggplot objects}
\usage{
\method{ggplot_add}{pwd_theme}(object, plot, object_name)
}
\value{
a modified ggplot2 object
}
\description{
Add PWD themes to ggplot objects
}
