% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\name{import_open_sans}
\alias{import_open_sans}
\title{Import the Open Sans Font}
\usage{
import_open_sans()
}
\description{
This function will import the open sans font package so that it
can be used in R plots. This font is licensed under the Apache
License Version 2.0 http://www.apache.org/licenses/
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
  font directory is displayed after the base import is complete. It is highly
  recommended that you install them on your system the same way you would any
  other font you wish to use in other programs.
}
