% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\name{import_public_sans}
\alias{import_public_sans}
\title{Import the Public Sans font}
\usage{
import_public_sans()
}
\description{
This function will import the public sans font to R for use in plots.
Public Sans is Copyright 2015 Impallari Type and licensed under the
SIL Open Font License, Version 1.1
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
  font directory is displayed after the base import is complete. It is highly
  recommended that you install them on your system the same way you would any
  other font you wish to use in other programs.
}
