% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\docType{data}
\name{OpenSans}
\alias{OpenSans}
\alias{font_os}
\alias{font_os_light}
\alias{font_os_extrabold}
\alias{font_os_semibold}
\title{Open Sans Font Variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_os

font_os_light

font_os_extrabold

font_os_semibold
}
\description{
font_os == "Open Sans"

font_os_light == "Open Sans Light"

font_os_extrabold == "Open Sans ExtraBold"

font_os_semibold == "Open Sans SemiBold"
}
\keyword{datasets}
