% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-fonts.R
\docType{data}
\name{PublicSans}
\alias{PublicSans}
\alias{font_pub}
\alias{font_pub_light}
\alias{font_pub_thin}
\alias{font_pub_black}
\alias{font_pub_extrabold}
\alias{font_pub_etralight}
\alias{font_pub_medium}
\alias{font_pub_semibold}
\title{Public Sans Font variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_pub

font_pub_light

font_pub_thin

font_pub_black

font_pub_extrabold

font_pub_etralight

font_pub_medium

font_pub_semibold
}
\description{
font_pub == "Public Sans"

font_pub_light == "Public Sans Light"

font_pub_thin == "Public Sans Thin"

font_pub_black == "Public Sans Black"

font_pub_extrabold == "Public Sans ExtraBold"

font_pub_extralight == "Public Sans ExtraLight"

font_pub_medium == "Public Sans Medium"

font_pub_semibold == "Public Sans SemiBold"
}
\keyword{datasets}
