% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{pwd_cols}
\alias{pwd_cols}
\title{Function to extract PWD colors as hex codes}
\usage{
pwd_cols(...)
}
\arguments{
\item{...}{Character names of pwd_colors}
}
\description{
Function to extract PWD colors as hex codes
}
