% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pwd.R
\name{pwd_geoms}
\alias{pwd_geoms}
\alias{undo_pwd_geoms}
\title{Replace the default color and fill options with PWD themed colors}
\usage{
pwd_geoms()

undo_pwd_geoms()
}
\description{
Replace the default color and fill options with PWD themed colors
}
\details{
This function will overwrite default color and fill options provided
by ggplot2 to match PWD themes more closely. By default, this function
will be run when loading the pwdthemes package. It can be easily undone
by running the undo_pwd_geoms function.
}
