% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{pwd_pals}
\alias{pwd_pals}
\alias{view_pwd_palette}
\title{Return function to interpolate a PWD color palette}
\usage{
pwd_pals(palette = "main", reverse = FALSE, ...)

view_pwd_palette(palette = "main", n = NULL, show_hex = TRUE)
}
\arguments{
\item{palette}{Character name of palette in pwd_palettes}

\item{reverse}{Logical, if TRUE the palette is reversed}

\item{...}{Additional arguments to pass to colorRampPalette}

\item{n}{number of desired levels in palette}

\item{show_hex}{logical, if TRUE (default), the hex code will be printed
for each color. If a large n is specified, it is recommended to set this
to false, as the hex codes become unreadable if n > 100}
}
\description{
Return function to interpolate a PWD color palette
}
\details{
This function will create a function to interpolate
different PWD color palettes. Its main argument, \code{palette},
takes any one of the following options: main, main_dark, main_light,
highlights, highlights_dark, highlights_light, map, map_dark,
map_light, dark, light, or full.

To see what these different palette types look like, use the
\code{view_pwd_palette(...)} function. You can pass the palette
you wish to view and the number of colors you want to see what
that palette will look like.
}
\section{Functions}{
\itemize{
\item \code{view_pwd_palette}: Return figure with selected PWD color palette
}}

