% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-styles.R
\name{pwd_styling}
\alias{pwd_styling}
\title{Add PWD styling to kable functions}
\usage{
pwd_styling(kable, options = c("striped", "bordered"), ...)
}
\arguments{
\item{kable}{the kable that you with to style}

\item{options}{the options to be passed to latex_options or bootstrap_options
arguments in \code{\link[kableExtra]{kable_extra}} depending on whether
the format is latex or html, respectively}

\item{...}{additional params to be passed to \code{\link[kableExtra]{kable_styling}}}
}
\description{
Add PWD styling to kable functions
}
\details{
This function will add styling to your kable tables used in Rmarkdown
documents. It will add the "striped" option to the table and make the
background of the column names PWD blue with white font. In pdf outputs
it will also change the background color of the striped rows. If you
are knitting to an HTML document than you will need to use the
\code{\link[pwdthemes]{set_pwd_styling}}.

It is important to note that if you use this function, than you must
pass all other arguments intended for \code{\link[kableExtra]{kable_styling}}
to this function. Currently, you can not use the kable_styling function
on a kable that it has previously been applied to.
}
