% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_pwd}
\alias{scale_color_pwd}
\alias{scale_colour_pwd}
\alias{scale_fill_pwd}
\title{Color scale constructor for PWD colors}
\usage{
scale_color_pwd(palette = "main", reverse = FALSE, discrete = TRUE,
  ...)

scale_colour_pwd(palette = "main", reverse = FALSE, discrete = TRUE,
  ...)

scale_fill_pwd(palette = "main", reverse = FALSE, discrete = TRUE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in pwd_palettes}

\item{reverse}{logical, if TRUE the palette is reversed}

\item{discrete}{logical, if TRUE (default) the color aesthetic
is treated as discrete}

\item{...}{Additional parameters to be passed to either
\code{ggplot2::discrete_scale} or \code{ggplot2::scale_color_gradientn},
depending on whehter discrete is TRUE or FALSE}
}
\description{
Color scale constructor for PWD colors
}
