% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-styles.R
\name{set_pwd_styling}
\alias{set_pwd_styling}
\title{Define CSS to format HTML kable}
\usage{
set_pwd_styling()
}
\description{
Define CSS to format HTML kable
}
\details{
This function will define css in your HTML Rmarkdown document
that will customize the format of kable's to meet PWD style
guidelines. This function only needs to be called once at the beginning
of your Rmarkdown document and will only work with html output.

The css makes the header of the table have a blue background with white
letters and the striped table have a light cyan background.
}
