% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-pwd.R
\name{theme_pwd}
\alias{theme_pwd}
\title{A ggplot2 theme for PWD}
\usage{
theme_pwd(base_size = 12, base_family = "Calibri")
}
\description{
This function will apply a custom PWD ggplot2 theme to
ggplot2 objects that matches the style guidelines
provided by the Public Affairs division
}
